+++ gltf.go
@@ -12,8 +12,8 @@ func Index(i uint32) *uint32 {
 	return &i
 }
 
-// Float is an utility function that returns a pointer to a float32.
-func Float(val float32) *float32 {
+// Float is an utility function that returns a pointer to a float64.
+func Float(val float64) *float64 {
 	return &val
 }
 
@@ -75,8 +75,8 @@ type Accessor struct {
 	Normalized    bool          `json:"normalized,omitempty"`      // Specifies whether integer data values should be normalized.
 	Count         uint32        `json:"count" validate:"required"` // The number of attributes referenced by this accessor.
 	Type          AccessorType  `json:"type" validate:"lte=6"`
-	Max           []float32     `json:"max,omitempty" validate:"omitempty,lte=16"` // Maximum value of each component in this attribute.
-	Min           []float32     `json:"min,omitempty" validate:"omitempty,lte=16"` // Minimum value of each component in this attribute.
+	Max           []float64     `json:"max,omitempty" validate:"omitempty,lte=16"` // Maximum value of each component in this attribute.
+	Min           []float64     `json:"min,omitempty" validate:"omitempty,lte=16"` // Minimum value of each component in this attribute.
 	Sparse        *Sparse       `json:"sparse,omitempty"`                          // Sparse storage of attributes that deviate from their initialization value.
 }
 
@@ -173,16 +173,16 @@ type Node struct {
 	Camera      *uint32     `json:"camera,omitempty"`
 	Children    []uint32    `json:"children,omitempty" validate:"omitempty,unique"`
 	Skin        *uint32     `json:"skin,omitempty"`
-	Matrix      [16]float32 `json:"matrix"` // A 4x4 transformation matrix stored in column-major order.
+	Matrix      [16]float64 `json:"matrix"` // A 4x4 transformation matrix stored in column-major order.
 	Mesh        *uint32     `json:"mesh,omitempty"`
-	Rotation    [4]float32  `json:"rotation" validate:"omitempty,dive,gte=-1,lte=1"` // The node's unit quaternion rotation in the order (x, y, z, w), where w is the scalar.
-	Scale       [3]float32  `json:"scale"`
-	Translation [3]float32  `json:"translation"`
-	Weights     []float32   `json:"weights,omitempty"` // The weights of the instantiated Morph Target.
+	Rotation    [4]float64  `json:"rotation" validate:"omitempty,dive,gte=-1,lte=1"` // The node's unit quaternion rotation in the order (x, y, z, w), where w is the scalar.
+	Scale       [3]float64  `json:"scale"`
+	Translation [3]float64  `json:"translation"`
+	Weights     []float64   `json:"weights,omitempty"` // The weights of the instantiated Morph Target.
 }
 
 // MatrixOrDefault returns the node matrix if it represents a valid affine matrix, else return the default one.
-func (n *Node) MatrixOrDefault() [16]float32 {
+func (n *Node) MatrixOrDefault() [16]float64 {
 	if n.Matrix == emptyMatrix {
 		return DefaultMatrix
 	}
@@ -190,7 +190,7 @@ func (n *Node) MatrixOrDefault() [16]float32 {
 }
 
 // RotationOrDefault returns the node rotation if it represents a valid quaternion, else return the default one.
-func (n *Node) RotationOrDefault() [4]float32 {
+func (n *Node) RotationOrDefault() [4]float64 {
 	if n.Rotation == emptyRotation {
 		return DefaultRotation
 	}
@@ -198,7 +198,7 @@ func (n *Node) RotationOrDefault() [4]float32 {
 }
 
 // ScaleOrDefault returns the node scale if it represents a valid scale factor, else return the default one.
-func (n *Node) ScaleOrDefault() [3]float32 {
+func (n *Node) ScaleOrDefault() [3]float64 {
 	if n.Scale == emptyScale {
 		return DefaultScale
 	}
@@ -206,7 +206,7 @@ func (n *Node) ScaleOrDefault() [3]float32 {
 }
 
 // TranslationOrDefault returns the node translation.
-func (n *Node) TranslationOrDefault() [3]float32 {
+func (n *Node) TranslationOrDefault() [3]float64 {
 	return n.Translation
 }
 
@@ -233,20 +233,20 @@ type Camera struct {
 type Orthographic struct {
 	Extensions Extensions  `json:"extensions,omitempty"`
 	Extras     interface{} `json:"extras,omitempty"`
-	Xmag       float32     `json:"xmag"`                               // The horizontal magnification of the view.
-	Ymag       float32     `json:"ymag"`                               // The vertical magnification of the view.
-	Zfar       float32     `json:"zfar" validate:"gt=0,gtfield=Znear"` // The distance to the far clipping plane.
-	Znear      float32     `json:"znear" validate:"gte=0"`             // The distance to the near clipping plane.
+	Xmag       float64     `json:"xmag"`                               // The horizontal magnification of the view.
+	Ymag       float64     `json:"ymag"`                               // The vertical magnification of the view.
+	Zfar       float64     `json:"zfar" validate:"gt=0,gtfield=Znear"` // The distance to the far clipping plane.
+	Znear      float64     `json:"znear" validate:"gte=0"`             // The distance to the near clipping plane.
 }
 
 // Perspective camera containing properties to create a perspective projection matrix.
 type Perspective struct {
 	Extensions  Extensions  `json:"extensions,omitempty"`
 	Extras      interface{} `json:"extras,omitempty"`
-	AspectRatio *float32    `json:"aspectRatio,omitempty"`
-	Yfov        float32     `json:"yfov"`           // The vertical field of view in radians.
-	Zfar        *float32    `json:"zfar,omitempty"` // The distance to the far clipping plane.
-	Znear       float32     `json:"znear"`          // The distance to the near clipping plane.
+	AspectRatio *float64    `json:"aspectRatio,omitempty"`
+	Yfov        float64     `json:"yfov"`           // The vertical field of view in radians.
+	Zfar        *float64    `json:"zfar,omitempty"` // The distance to the far clipping plane.
+	Znear       float64     `json:"znear"`          // The distance to the near clipping plane.
 }
 
 // A Mesh is a set of primitives to be rendered. A node can contain one mesh. A node's transform places the mesh in the scene.
@@ -255,7 +255,7 @@ type Mesh struct {
 	Extras     interface{}  `json:"extras,omitempty"`
 	Name       string       `json:"name,omitempty"`
 	Primitives []*Primitive `json:"primitives" validate:"required,gt=0,dive"`
-	Weights    []float32    `json:"weights,omitempty"`
+	Weights    []float64    `json:"weights,omitempty"`
 }
 
 // Primitive defines the geometry to be rendered with the given material.
@@ -278,14 +278,14 @@ type Material struct {
 	NormalTexture        *NormalTexture        `json:"normalTexture,omitempty"`
 	OcclusionTexture     *OcclusionTexture     `json:"occlusionTexture,omitempty"`
 	EmissiveTexture      *TextureInfo          `json:"emissiveTexture,omitempty"`
-	EmissiveFactor       [3]float32            `json:"emissiveFactor,omitempty" validate:"dive,gte=0,lte=1"`
+	EmissiveFactor       [3]float64            `json:"emissiveFactor,omitempty" validate:"dive,gte=0,lte=1"`
 	AlphaMode            AlphaMode             `json:"alphaMode,omitempty" validate:"lte=2"`
-	AlphaCutoff          *float32              `json:"alphaCutoff,omitempty" validate:"omitempty,gte=0"`
+	AlphaCutoff          *float64              `json:"alphaCutoff,omitempty" validate:"omitempty,gte=0"`
 	DoubleSided          bool                  `json:"doubleSided,omitempty"`
 }
 
 // AlphaCutoffOrDefault returns the scale if it is not nil, else return the default one.
-func (m *Material) AlphaCutoffOrDefault() float32 {
+func (m *Material) AlphaCutoffOrDefault() float64 {
 	if m.AlphaCutoff == nil {
 		return 0.5
 	}
@@ -298,11 +298,11 @@ type NormalTexture struct {
 	Extras     interface{} `json:"extras,omitempty"`
 	Index      *uint32     `json:"index,omitempty"`
 	TexCoord   uint32      `json:"texCoord,omitempty"` // The index of texture's TEXCOORD attribute used for texture coordinate mapping.
-	Scale      *float32    `json:"scale,omitempty"`
+	Scale      *float64    `json:"scale,omitempty"`
 }
 
 // ScaleOrDefault returns the scale if it is not nil, else return the default one.
-func (n *NormalTexture) ScaleOrDefault() float32 {
+func (n *NormalTexture) ScaleOrDefault() float64 {
 	if n.Scale == nil {
 		return 1
 	}
@@ -315,11 +315,11 @@ type OcclusionTexture struct {
 	Extras     interface{} `json:"extras,omitempty"`
 	Index      *uint32     `json:"index,omitempty"`
 	TexCoord   uint32      `json:"texCoord,omitempty"` // The index of texture's TEXCOORD attribute used for texture coordinate mapping.
-	Strength   *float32    `json:"strength,omitempty" validate:"omitempty,gte=0,lte=1"`
+	Strength   *float64    `json:"strength,omitempty" validate:"omitempty,gte=0,lte=1"`
 }
 
 // StrengthOrDefault returns the strength if it is not nil, else return the default one.
-func (o *OcclusionTexture) StrengthOrDefault() float32 {
+func (o *OcclusionTexture) StrengthOrDefault() float64 {
 	if o.Strength == nil {
 		return 1
 	}
@@ -330,15 +330,15 @@ func (o *OcclusionTexture) StrengthOrDefault() float32 {
 type PBRMetallicRoughness struct {
 	Extensions               Extensions   `json:"extensions,omitempty"`
 	Extras                   interface{}  `json:"extras,omitempty"`
-	BaseColorFactor          *[4]float32  `json:"baseColorFactor,omitempty" validate:"omitempty,dive,gte=0,lte=1"`
+	BaseColorFactor          *[4]float64  `json:"baseColorFactor,omitempty" validate:"omitempty,dive,gte=0,lte=1"`
 	BaseColorTexture         *TextureInfo `json:"baseColorTexture,omitempty"`
-	MetallicFactor           *float32     `json:"metallicFactor,omitempty" validate:"omitempty,gte=0,lte=1"`
-	RoughnessFactor          *float32     `json:"roughnessFactor,omitempty" validate:"omitempty,gte=0,lte=1"`
+	MetallicFactor           *float64     `json:"metallicFactor,omitempty" validate:"omitempty,gte=0,lte=1"`
+	RoughnessFactor          *float64     `json:"roughnessFactor,omitempty" validate:"omitempty,gte=0,lte=1"`
 	MetallicRoughnessTexture *TextureInfo `json:"metallicRoughnessTexture,omitempty"`
 }
 
 // MetallicFactorOrDefault returns the metallic factor if it is not nil, else return the default one.
-func (p *PBRMetallicRoughness) MetallicFactorOrDefault() float32 {
+func (p *PBRMetallicRoughness) MetallicFactorOrDefault() float64 {
 	if p.MetallicFactor == nil {
 		return 1
 	}
@@ -346,7 +346,7 @@ func (p *PBRMetallicRoughness) MetallicFactorOrDefault() float32 {
 }
 
 // RoughnessFactorOrDefault returns the roughness factor if it is not nil, else return the default one.
-func (p *PBRMetallicRoughness) RoughnessFactorOrDefault() float32 {
+func (p *PBRMetallicRoughness) RoughnessFactorOrDefault() float64 {
 	if p.RoughnessFactor == nil {
 		return 1
 	}
@@ -354,9 +354,9 @@ func (p *PBRMetallicRoughness) RoughnessFactorOrDefault() float32 {
 }
 
 // BaseColorFactorOrDefault returns the base color factor if it is not nil, else return the default one.
-func (p *PBRMetallicRoughness) BaseColorFactorOrDefault() [4]float32 {
+func (p *PBRMetallicRoughness) BaseColorFactorOrDefault() [4]float64 {
 	if p.BaseColorFactor == nil {
-		return [4]float32{1, 1, 1, 1}
+		return [4]float64{1, 1, 1, 1}
 	}
 	return *p.BaseColorFactor
 }
